requirements.txt
numpy~=1.22.3
scikit-learn~=1.0.2
scipy~=1.8.0
matplotlib~=3.5.1
joblib~=1.1.0
PyYAML